+++ source/MaterialXView/Viewer.h
@@ -21,7 +21,6 @@ class Viewer : public ng::Screen
            const DocumentModifiers& modifiers,
            mx::HwSpecularEnvironmentMethod specularEnvironmentMethod,
            const std::string& envRadiancePath,
-           const std::string& envIrradiancePath,
            int multiSampleCount);
     ~Viewer() { }
 
@@ -86,7 +85,8 @@ class Viewer : public ng::Screen
     void drawScene3D();
     void drawScene2D();
 
-    void setupLights(mx::DocumentPtr doc);
+    void loadEnvironmentLight();
+    void loadDirectLights(mx::DocumentPtr doc);
     void loadDocument(const mx::FilePath& filename, mx::DocumentPtr libraries);
     void reloadShaders();
     void loadStandardLibraries();
@@ -109,6 +109,7 @@ class Viewer : public ng::Screen
 
     void createLoadMeshInterface(Widget* parent, const std::string& label);
     void createLoadMaterialsInterface(Widget* parent, const std::string& label);
+    void createLoadEnvironmentInterface(Widget* parent, const std::string& label);
     void createSaveMaterialsInterface(Widget* parent, const std::string& label);
     void createPropertyEditorInterface(Widget* parent, const std::string& label);
     void createAdvancedSettings(Widget* parent);
@@ -146,15 +147,14 @@ class Viewer : public ng::Screen
     mx::StringSet _xincludeFiles;
 
     // Lighting information
-    std::string _lightFileName;
-    std::string _envRadiancePath;
-    std::string _envIrradiancePath;
+    mx::FilePath _lightFileName;
+    mx::FilePath _envRadiancePath;
     bool _directLighting;
     bool _indirectLighting;
     bool _ambientOcclusion;
 
     // Geometry selections
-    std::string _meshFilename;
+    mx::FilePath _meshFilename;
     std::vector<mx::MeshPartitionPtr> _geometryList;
     size_t _selectedGeom;
     ng::Label* _geomLabel;