+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.36.5] - Development
 
 ### Added
+- Added a Load Environment option to the viewer.
 - Added an initial TextureBaker class, supporting baking of graph outputs to textures.
 - Added initial support for units, including the MaterialX\:\:Unit, MaterialX\:\:UnitDef, and MaterialX\:\:UnitTypeDef classes.
 - Added support for unit conversion in shader code generation.